refactor: comment 연관관계 해제로 인한 쿼리 수정 및 개선

+++ src/main/java/com/depromeet/stonebed/domain/feed/dto/FindFeedDto.java
@@ -8,13 +8,13 @@ public record FindFeedDto(
         Mission mission,
         MissionRecord missionRecord,
         Member author,
-        Integer totalCommentCount,
+        Long totalCommentCount,
         Long totalBoostCount) {
     public static FindFeedDto from(
             Mission mission,
             MissionRecord missionRecord,
             Member author,
-            Integer totalCommentCount,
+            Long totalCommentCount,
             Long totalBoostCount) {
         return new FindFeedDto(mission, missionRecord, author, totalCommentCount, totalBoostCount);
     }
